+++ test/with-typescript/index.ts
@@ -111,6 +111,7 @@ let sessionConfig: SessionTypeInput = {
                 getAccessTokenLifeTimeMS: originalImpl.getAccessTokenLifeTimeMS,
                 getRefreshTokenLifeTimeMS: originalImpl.getRefreshTokenLifeTimeMS,
                 getSessionInformation: originalImpl.getSessionInformation,
+                regenerateAccessToken: originalImpl.regenerateAccessToken,
             };
         },
     },